redacted: main.c lab1.c lab1.h  int >> size_t

+++ lab1.c
@@ -5,7 +5,7 @@
 // --- 1 part ---
 
 void recursiv_sort(int* mass, int low, int high);
-void reverse_mass(int* mass, int size);
+void reverse_mass(int* mass, size_t size);
 
 void swap(int* a, int* b)
 {
@@ -14,9 +14,9 @@ void swap(int* a, int* b)
     *a = *a - *b;
 }
 
-void sort_mass(int* mass, const int size, const enum ESort eSort)
+void sort_mass(int* mass, const size_t size, const enum ESort eSort)
 {
-    recursiv_sort(mass, 0, size-1);
+    recursiv_sort(mass, 0, (int)size-1);
     if (eSort)
     {
         reverse_mass(mass, size);
@@ -56,22 +56,22 @@ void recursiv_sort(int* mass, int low, int high)
     recursiv_sort(mass, pointer, high);
 }
 
-void reverse_mass(int* mass, int size)
+void reverse_mass(int* mass, size_t size)
 {
-    for (int i = 0; i < size/2; i++)
+    for (int i = 0; i < (int)size/2; i++)
     {
-        swap(&mass[i], &mass[size-i-1]);
+        swap(&mass[i], &mass[(int)size-i-1]);
     }
 }
 
 // --- 2 part ---
 
-int get_column_len(const int* mass, const int size, const unsigned int column, int numb)
+int get_column_len(const int* mass, const size_t size, const unsigned int column, int numb)
 {
     int len = 0;
     int cur_len = 0;
     int k = 0; // buffer
-    for (int i = numb; i < size; i += column)
+    for (int i = numb; i < (int)size; i += column)
     {
         cur_len = 1;
         k = mass[i] / 10;
@@ -88,17 +88,17 @@ int get_column_len(const int* mass, const int size, const unsigned int column, i
     return len;
 }
 
-void print(const int* mass, const int size, const unsigned int column)
+void print(const int* mass, const size_t size, const unsigned int column)
 {
     int column_len[column];
     for (int i = 0; i < (int)column; i++)
     {
         column_len[i] = get_column_len(mass, size, column, i);
     }
-    for (int i = 0; i < size; i++)
+    for (int i = 0; i < (int)size; i++)
     {
         printf("%+-*d ", column_len[i % column] + 1, mass[i]);
-        if ((i % column == column - 1) || (i == size - 1))
+        if ((i % column == column - 1) || (i == (int)size - 1))
         {
             printf("\n");
         }
@@ -107,39 +107,39 @@ void print(const int* mass, const int size, const unsigned int column)
 
 // --- 3 part ---
 
-float get_median(const int* mass, const int size)
+float get_median(const int* mass, const size_t size)
 {
-    int mass2[size];
-    for (int i = 0; i < size; i++)
+    int mass2[(int)size];
+    for (int i = 0; i < (int)size; i++)
     {
         mass2[i] = mass[i];
     }
     sort_mass(mass2, size, E_UP);
-    if (size % 2 == 1)
+    if ((int)size % 2 == 1)
     {
-        return (float)mass2[size/2];
+        return (float)mass2[(int)size/2];
     }
     else
     {
-        return (float)(((float)(mass2[size/2-1]) + (float)(mass2[size/2]))/2);
+        return (float)(((float)(mass2[(int)size/2-1]) + (float)(mass2[(int)size/2]))/2);
     }
 }
 
 // --- 4 part ---
 
-void print_moda(const int* mass, const int size)
+void print_moda(const int* mass, const size_t size)
 {
-    int mass2[size];
-    for (int i = 0; i < size; i++)
+    int mass2[(int)size];
+    for (int i = 0; i < (int)size; i++)
     {
         mass2[i] = mass[i];
     }
     sort_mass(mass2, size, E_UP);
-    int mass_counter[size];
+    int mass_counter[(int)size];
     int cur_val = mass2[0];
     int cur_moda = 0;
     int moda = 0;
-    for (int i = 0; i < size; i++)
+    for (int i = 0; i < (int)size; i++)
     {
         if (mass2[i] == cur_val)
         {
@@ -156,7 +156,7 @@ void print_moda(const int* mass, const int size)
             mass_counter[i] = 1;
         }
     }
-    for (int i = 0; i < size; i++)
+    for (int i = 0; i < (int)size; i++)
     {
         if (mass_counter[i] == moda)
         {
@@ -168,10 +168,10 @@ void print_moda(const int* mass, const int size)
 
 // --- 5 part ---
 
-float get_average(const int* mass, const int size)
+float get_average(const int* mass, const size_t size)
 {
     float summ = 0;
-    for (int i = 0; i < size; i++)
+    for (int i = 0; i < (int)size; i++)
     {
         summ += (float)(mass[i]);
     }
@@ -180,10 +180,10 @@ float get_average(const int* mass, const int size)
 
 // --- 6 part ---
 
-int get_max(const int* mass, const int size)
+int get_max(const int* mass, const size_t size)
 {
     int m = mass[0];
-    for (int i = 0; i < size; i++)
+    for (int i = 0; i < (int)size; i++)
     {
         if (mass[i] > m)
         {
@@ -195,10 +195,10 @@ int get_max(const int* mass, const int size)
 
 // --- 7 part ---
 
-int get_min(const int* mass, const int size)
+int get_min(const int* mass, const size_t size)
 {
     int m = mass[0];
-    for (int i = 0; i < size; i++)
+    for (int i = 0; i < (int)size; i++)
     {
         if (mass[i] < m)
         {
@@ -210,27 +210,27 @@ int get_min(const int* mass, const int size)
 
 // --- 8 part ---
 
-int get_range(const int* mass, const int size)
+int get_range(const int* mass, const size_t size)
 {
     return get_max(mass, size) - get_min(mass, size);
 }
 
 // --- 9 part ---
 
-float get_dispersion(const int* mass, const int size)
+float get_dispersion(const int* mass, const size_t size)
 {
     float avg = get_average(mass, size);
     float sm = 0;
-    for (int i = 0; i < size; i++)
+    for (int i = 0; i < (int)size; i++)
     {
         sm += ((float)mass[i] - avg) * ((float)mass[i] - avg);
     }
-    return sm / (float)(size-1);
+    return sm / (float)((int)size - 1);
 }
 
 // --- 10 part ---
 
-float get_st_deviation(const int* mass, const int size)
+float get_st_deviation(const int* mass, const size_t size)
 {
     return pow(get_dispersion(mass, size), 0.5f);
 }
@@ -248,24 +248,24 @@ int count_steps(float a, float step)
     return s;
 }
 
-void hist(const int* mass, const int size, int* count, const int bins)
+void hist(const int* mass, const size_t size, int* count, const size_t bins)
 {
     int mn = get_min(mass, size);
     int mx = get_max(mass, size);
     float step = (float)(mx - mn) / (float)(bins);
-    for (int i = 0; i < bins; i++)
+    for (int i = 0; i < (int)bins; i++)
     {
         count[i] = 0;
     }
     float k = 0; // buffer
     int c_steps = 0; // buffer
-    for (int i = 0; i < size; i++)
+    for (int i = 0; i < (int)size; i++)
     {
         k = (float)(mass[i]) - (float)mn;
         c_steps = count_steps(k, step);
-        if (c_steps == bins)
+        if (c_steps == (int)bins)
         {
-            count[bins - 1] += 1;
+            count[(int)bins - 1] += 1;
         }
         else
         {
@@ -276,16 +276,16 @@ void hist(const int* mass, const int size, int* count, const int bins)
 
 // --- 12 part ---
 
-void print_hist(const int* mass, const int size, const int* count, const int bins)
+void print_hist(const int* mass, const size_t size, const int* count, const size_t bins)
 {
     int mn = get_min(mass, size);
     int mx = get_max(mass, size);
     float step = (float)(mx-mn) / (float)bins;
     int buffer[2] = {mn, mx}; // buffer
     int sdv = get_column_len(buffer, 2, 1, 0) + 8; // column correction
-    for (int i = 0; i < bins; i++)
+    for (int i = 0; i < (int)bins; i++)
     {
-        if (i < bins-1)
+        if (i < (int)bins-1)
         {
             printf("[ %*f, %*f) ", sdv, mn + (step*i), sdv, mn + (step*(i+1)));
         }
