+++ main.c
@@ -3,9 +3,9 @@
 #include <time.h>
 #include "lab1.h"
 
-const int size = 100;
+const size_t size = 100;
 const unsigned int column = 10;
-const int bins = 10;
+const size_t bins = 10;
 
 int main()
 {
@@ -22,7 +22,7 @@ int main()
     }
     int mass[size];
     srand(time(0));
-    for (int i = 0; i < size; i++)
+    for (int i = 0; i < (int)size; i++)
     {
         mass[i] = (rand()%(b-a+1))+a;
     }
@@ -37,7 +37,7 @@ int main()
     printf("%d\n", get_range(mass, size));
     printf("%f\n", get_dispersion(mass, size));
     printf("%f\n", get_st_deviation(mass, size));
-    int count[bins];
+    int count[(int)bins];
     hist(mass, size, count, bins);
     print_hist(mass, size, count, bins);
     return 0;
